Types: use type imports to remove references when building for production
