+++ packages/connect-core/src/connections/IOrganizationConnector.ts
@@ -1,4 +1,4 @@
-import {
+import type {
   AppFilters,
   Network,
   SubscriptionCallback,
