+++ packages/connect-core/src/entities/Organization.ts
@@ -1,4 +1,4 @@
-import {
+import type {
   Address,
   AppFilters,
   AppFiltersParam,
