+++ packages/connect-core/src/entities/Transaction.ts
@@ -1,4 +1,4 @@
-import { Address } from '@1hive/connect-types'
+import type { Address } from '@1hive/connect-types'
 
 import { TransactionData } from '../types'
