+++ packages/connect-core/src/utils/app-connectors.ts
@@ -1,13 +1,16 @@
-import {
+import type {
   Network,
   SubscriptionCallback,
   SubscriptionHandler,
   SubscriptionResult,
   SubscriptionStart,
+} from '@1hive/connect-types'
+import {
   isSubscriptionHandler,
   isSubscriptionStart,
-} from '@1hive/connect-types'
-import { ConnectionContext } from '../types'
+  ConnectionContext,
+} from '../types'
+
 import { ErrorInvalidConnector, ErrorInvalidApp } from '../errors'
 import App from '../entities/App'
