+++ packages/connect-ethereum/src/index.ts
@@ -3,7 +3,7 @@
 /* eslint-disable @typescript-eslint/no-empty-function */
 /* eslint-disable @typescript-eslint/explicit-function-return-type */
 
-import {
+import type {
   Address,
   AppFilters,
   Network,
