--- packages/connect-react/connect-ethereum/.yarnrc.yml
@@ -1,1 +1,0 @@
-yarnPath: .yarn/releases/yarn-1.22.4.js
