--- packages/connect-react/connect-ethereum/jest.config.js
@@ -1,14 +1,0 @@
-const baseConfig = require('../../jest.config.base')
-
-const packageName = require('./package.json').name.split('@1hive/').pop()
-
-module.exports = {
-  ...baseConfig,
-  roots: [`<rootDir>/packages/${packageName}`],
-  collectCoverageFrom: ['src/**/*.{ts,tsx}'],
-  testRegex: `(packages/${packageName}/.*/__tests__/.*|\\.(test|spec))\\.tsx?$`,
-  modulePaths: [`<rootDir>/packages/${packageName}/src/`],
-  name: packageName,
-  displayName: 'ETHEREUM',
-  rootDir: '../..',
-}
