--- packages/connect-react/connect-ethereum/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "outDir": "dist/esm",
-    "rootDir": "src"
-  },
-  "include": ["src"],
-  "exclude": ["dist", "**/__test__/**"],
-  "references": [{ "path": "../connect-core" }]
-}
