--- packages/connect-thegraph/connect-ethereum/package.json
@@ -1,51 +1,0 @@
-{
-  "name": "@1hive/connect-ethereum",
-  "version": "0.8.0-alpha.6",
-  "license": "LGPL-3.0-or-later",
-  "main": "dist/cjs/index.js",
-  "module": "dist/esm/index.js",
-  "types": "dist/cjs/index.d.ts",
-  "sideEffects": false,
-  "scripts": {
-    "build": "yarn build:esm && yarn build:cjs",
-    "build:esm": "tsc --build",
-    "build:cjs": "tsc --build ./tsconfig.cjs.json",
-    "dev:esm": "yarn build:esm --watch",
-    "dev:cjs": "yarn build:cjs --watch",
-    "test": "is-ci \"test:coverage\" \"test:watch\"",
-    "test:coverage": "jest --coverage",
-    "test:watch": "jest --watch",
-    "test:debug": "node --inspect-brk ../../node_modules/jest/bin/jest.js --runInBand --watch",
-    "clean": "rm -rf ./dist && rm -f tsconfig.tsbuildinfo",
-    "lint": "eslint --ext .ts ./src"
-  },
-  "files": [
-    "dist"
-  ],
-  "devDependencies": {
-    "@types/jest": "^26.0.0",
-    "@typescript-eslint/eslint-plugin": "^2.29.0",
-    "@typescript-eslint/parser": "^2.29.0",
-    "eslint": "^6.8.0",
-    "is-ci-cli": "^2.1.1",
-    "jest": "^26.0.1",
-    "prettier": "^2.0.5",
-    "ts-jest": "^26.1.0",
-    "typescript": "^3.8.3"
-  },
-  "dependencies": {
-    "@1hive/connect-core": "^0.8.0-alpha.6"
-  },
-  "description": "Access and interact with Aragon Organizations and their apps.",
-  "keywords": [
-    "ethereum",
-    "javascript",
-    "web3",
-    "dapp",
-    "aragon"
-  ],
-  "author": "Aragon Association <legal@1hive.org>",
-  "homepage": "https://github.com/aragon/connect",
-  "bugs": "https://github.com/aragon/connect/issues",
-  "repository": "https://github.com/aragon/connect"
-}
